chore(nestjs-trpc): refactored util files and functions
