+++ packages/nestjs-trpc/lib/decorators/router.decorator.ts
@@ -1,6 +1,8 @@
 import { applyDecorators, SetMetadata } from '@nestjs/common';
 import { ROUTER_METADATA_KEY } from '../trpc.constants';
-import { getCallerFilePath } from '../utils/path.utils';
+import { FileScanner } from '../scanners/file.scanner';
+
+const fileScanner = new FileScanner();
 
 /**
  * Decorator that marks a class as a TRPC router that can receive inbound
@@ -19,19 +21,8 @@ import { getCallerFilePath } from '../utils/path.utils';
  * @publicApi
  */
 export function Router(args?: { alias?: string }): ClassDecorator {
-  const stack = new Error().stack || '';
-  const stackLines = stack.split('\n');
-  const callerLine = stackLines[2]; // The line of the stack trace where this decorator is used
-
-  // Regex to extract the file path from the stack trace line
-  const match = callerLine.match(/\((.*):\d+:\d+\)$/);
-  const path = match ? match[1] : 'unknown';
-  const newPath = getCallerFilePath();
-
-  console.log({ originalPath: path, newPath });
+  const path = fileScanner.getCallerFilePath();
   return applyDecorators(
-    ...[
-      SetMetadata(ROUTER_METADATA_KEY, { alias: args?.alias, path: newPath }),
-    ],
+    ...[SetMetadata(ROUTER_METADATA_KEY, { alias: args?.alias, path })],
   );
 }
