+++ packages/nestjs-trpc/lib/generators/decorator.generator.ts
@@ -1,12 +1,21 @@
-import { Decorator, Project, SourceFile } from 'ts-morph';
+import {
+  Decorator,
+  Expression,
+  Project,
+  SourceFile,
+  SyntaxKind,
+} from 'ts-morph';
 import { DecoratorGeneratorMetadata } from '../interfaces/generator.interface';
-import { getDecoratorPropertyValue } from '../utils/file.util';
 import { ConsoleLogger, Inject, Injectable } from '@nestjs/common';
+import { ProcedureGenerator } from './procedure.generator';
 
 @Injectable()
 export class DecoratorGenerator {
   @Inject(ConsoleLogger)
   private readonly consoleLogger!: ConsoleLogger;
+
+  @Inject(ProcedureGenerator)
+  private readonly procedureGenerator!: ProcedureGenerator;
 
   public serializeProcedureDecorators(
     decorators: Decorator[],
@@ -18,13 +27,13 @@ export class DecoratorGenerator {
         const decoratorName = decorator.getName();
 
         if (decoratorName === 'Query' || decoratorName === 'Mutation') {
-          const input = getDecoratorPropertyValue(
+          const input = this.getDecoratorPropertyValue(
             decorator,
             'input',
             sourceFile,
             project,
           );
-          const output = getDecoratorPropertyValue(
+          const output = this.getDecoratorPropertyValue(
             decorator,
             'output',
             sourceFile,
@@ -48,5 +57,37 @@ export class DecoratorGenerator {
       },
       [],
     );
+  }
+
+  public getDecoratorPropertyValue(
+    decorator: Decorator,
+    propertyName: string,
+    sourceFile: SourceFile,
+    project: Project,
+  ): string | null {
+    const args = decorator.getArguments();
+
+    for (const arg of args) {
+      if (arg.getKind() === SyntaxKind.ObjectLiteralExpression) {
+        const properties = (arg as any).getProperties();
+        const property = properties.find(
+          (p: any) => p.getName() === propertyName,
+        );
+
+        if (!property) {
+          return null;
+        }
+
+        const propertyInitializer: Expression = property.getInitializer();
+        return this.procedureGenerator.flattenZodSchema(
+          propertyInitializer,
+          sourceFile,
+          project,
+          propertyInitializer.getText(),
+        );
+      }
+    }
+
+    return null;
   }
 }
