+++ packages/nestjs-trpc/lib/interfaces/scanner.interface.ts
@@ -1,0 +1,7 @@
+export interface SourceMapping {
+  version: number;
+  file: string;
+  sourceRoot: string;
+  sources: Array<string>;
+  mappings: string;
+}
