+++ packages/nestjs-trpc/lib/utils/ts-morph.util.ts
@@ -1,0 +1,8 @@
+import { SourceFile } from 'ts-morph';
+
+export async function saveOrOverrideFile(
+  sourceFile: SourceFile,
+): Promise<void> {
+  sourceFile.formatText({ indentSize: 2 });
+  await sourceFile.save();
+}
